+++ src/main/java/com/github/llytho/lootjs/loot/action/ReplaceLootAction.java
@@ -1,6 +1,7 @@
 package com.github.llytho.lootjs.loot.action;
 
 import com.github.llytho.lootjs.core.ILootAction;
+import com.github.llytho.lootjs.core.LootEntry;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.storage.loot.LootContext;
 
@@ -9,18 +10,25 @@ import java.util.function.Predicate;
 
 public class ReplaceLootAction implements ILootAction {
     private final Predicate<ItemStack> predicate;
-    private final ItemStack itemStack;
+    private final LootEntry lootEntry;
+    private final boolean preserveCount;
 
-    public ReplaceLootAction(Predicate<ItemStack> predicate, ItemStack itemStack) {
+    public ReplaceLootAction(Predicate<ItemStack> predicate, LootEntry lootEntry, boolean preserveCount) {
         this.predicate = predicate;
-        this.itemStack = itemStack;
+        this.lootEntry = lootEntry;
+        this.preserveCount = preserveCount;
     }
 
     @Override
     public boolean applyLootHandler(LootContext context, List<ItemStack> loot) {
         for (int i = 0; i < loot.size(); i++) {
-            if (predicate.test(loot.get(i))) {
-                loot.set(i, itemStack.copy());
+            ItemStack currentItemStack = loot.get(i);
+            if (predicate.test(currentItemStack)) {
+                ItemStack copy = this.lootEntry.apply(context);
+                if(preserveCount){
+                    copy.setCount(Math.min(currentItemStack.getCount(), copy.getMaxStackSize()));
+                }
+                loot.set(i, copy);
             }
         }